+++ cloudnet/src/main/java/de/dytanic/cloudnet/network/listener/PacketServerDatabaseAction.java
@@ -129,7 +129,7 @@ public class PacketServerDatabaseAction implements IPacketListener {
                     return;
                 }
                 if (message.equals("documentsCount")) {
-                    int documentsCount = database.getDocumentsCount();
+                    long documentsCount = database.getDocumentsCount();
                     this.sendResponse(channel, packet.getUniqueId(), new JsonDocument().append("documentsCount", documentsCount));
                     return;
                 }
